Move `design/Onboard` to `teleport/components/Onboard`
