+++ web/packages/teleport/src/Welcome/NewCredentials/Expired.tsx
@@ -20,7 +20,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { H2, P1 } from 'design';
 
-import { OnboardCard } from 'design/Onboard/OnboardCard';
+import { OnboardCard } from 'teleport/components/Onboard';
 
 export function Expired({ resetMode = false }) {
   const titleCodeTxt = resetMode ? 'Reset' : 'Invitation';
