+++ web/packages/teleport/src/Welcome/NewCredentials/NewCredentials.tsx
@@ -18,11 +18,9 @@
 
 import React, { useState } from 'react';
 import { NewFlow, StepSlider } from 'design/StepSlider';
-
-import { OnboardCard } from 'design/Onboard/OnboardCard';
-
 import { Box } from 'design';
 
+import { OnboardCard } from 'teleport/components/Onboard';
 import OSSRecoveryCodes from 'teleport/components/RecoveryCodes';
 import cfg from 'teleport/config';
