+++ web/packages/teleport/src/Welcome/NewCredentials/NewPasswordlessDevice.tsx
@@ -24,7 +24,7 @@ import Validation, { Validator } from 'shared/components/Validation';
 import { requiredField } from 'shared/components/Validation/rules';
 import { useRefAutoFocus } from 'shared/hooks';
 
-import { OnboardCard } from 'design/Onboard/OnboardCard';
+import { OnboardCard } from 'teleport/components/Onboard';
 
 import { PasskeyBlurb } from 'teleport/components/Passkeys/PasskeyBlurb';
