+++ web/packages/teleport/src/Welcome/Welcome.story.tsx
@@ -19,7 +19,7 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 
-import { WelcomeWrapper } from 'design/Onboard/WelcomeWrapper';
+import { WelcomeWrapper } from 'teleport/components/Onboard';
 
 import { NewCredentials } from 'teleport/Welcome/NewCredentials';
