+++ web/packages/teleport/src/components/Onboard/OnboardCard.ts
