+++ web/packages/teleport/src/components/Onboard/OnboardFooter.test.tsx
@@ -19,7 +19,8 @@
 import React from 'react';
 
 import { render, screen } from 'design/utils/testing';
-import { OnboardFooter } from 'design/Onboard/OnboardFooter';
+
+import { OnboardFooter } from './OnboardFooter';
 
 test('renders RR, TOS, and PP', () => {
   render(<OnboardFooter />);
