+++ web/packages/teleport/src/components/Onboard/OnboardFooter.tsx
