+++ web/packages/teleport/src/components/Onboard/WelcomeWrapper.tsx
@@ -20,10 +20,10 @@ import React from 'react';
 
 import styled from 'styled-components';
 
-import { LogoHero } from 'teleport/components/LogoHero';
-
 import { Box, Flex } from 'design';
 import cloudCity from 'design/assets/images/backgrounds/cloud-city.png';
+
+import { LogoHero } from 'teleport/components/LogoHero';
 
 import { OnboardFooter } from './OnboardFooter';
